refactor(workspace): make the project a workspace
Include in the workspace the astarte-device-sdk-derive package and
separate the e2e-test in a separated package with it's own dependencies.
Specify the common package field in the workspace.
