+++ .github/workflows/test.yaml
@@ -47,10 +47,10 @@ jobs:
         if: hashFiles('Cargo.lock') == ''
         run: cargo generate-lockfile
       - name: Run cargo test --locked
-        run: cargo test --locked --all-features --all-targets
+        run: cargo test --locked --all-features --all-targets --workspace
       # https://github.com/rust-lang/cargo/issues/6669
       - name: Run cargo test --doc
-        run: cargo test --locked --all-features --doc
+        run: cargo test --locked --all-features --doc --workspace
   os-check:
     runs-on: ${{ matrix.os }}
     name: ${{ matrix.os }} / stable / vcpkg
@@ -77,7 +77,7 @@ jobs:
         if: hashFiles('Cargo.lock') == ''
         run: cargo generate-lockfile
       - name: cargo test
-        run: cargo test --locked --all-features --all-targets
+        run: cargo test --locked --all-features --all-targets --workspace
   code_coverage_tarpauline:
     name: Code coverage tarpauline based
     runs-on: ubuntu-latest
