+++ astarte-device-sdk-derive/Cargo.toml
@@ -6,16 +6,18 @@
 
 [package]
 name = "astarte-device-sdk-derive"
-version = "0.6.1"
-edition = "2021"
-license = "Apache-2.0"
-repository = "https://github.com/astarte-platform/astarte-device-sdk-rust"
-description ="Derive macros implementation used by Astarte Device SDK"
-homepage = "https://astarte.cloud/"
+version = { workspace = true }
+categories = ["embedded", "api-bindings", "development-tools::procedural-macro-helpers"]
 documentation = "https://docs.rs/astarte-device-sdk-derive"
-readme ="README.md"
+edition = { workspace = true }
+exclude = { workspace = true }
+homepage = { workspace = true }
 keywords = ["sdk", "iot", "astarte", "derive"]
-rust-version = "1.66.1"
+license = { workspace = true }
+readme = "README.md"
+repository = { workspace = true }
+rust-version = { workspace = true }
+description = "Derive macros implementation used by Astarte Device SDK"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -23,6 +25,6 @@ rust-version = "1.66.1"
 proc-macro = true
 
 [dependencies]
-syn = {version = "2.0", features = ["full"]}
-quote = "1.0"
-proc-macro2 = "1.0.66"
+proc-macro2 = { workspace = true }
+quote = { workspace = true }
+syn = { workspace = true, features = ["full"] }
