+++ e2e-test/interfaces/org.astarte-platform.rust.e2etest.DeviceAggregate.json
